+++ weblate/trans/models/component.py
@@ -268,11 +268,11 @@ class ComponentQuerySet(models.QuerySet):
             "pull_message",
         )
 
-    def get_linked(self, val):
-        """Return component for linked repo."""
-        if not is_repo_link(val):
-            return None
-        project, *categories, component = val[10:].split("/")
+    def filter_by_path(self, path: str) -> ComponentQuerySet:
+        try:
+            project, *categories, component = path.split("/")
+        except ValueError:
+            raise Component.DoesNotExist
         kwargs = {}
         prefix = ""
         for category in reversed(categories):
@@ -280,7 +280,18 @@ class ComponentQuerySet(models.QuerySet):
             prefix = f"category__{prefix}"
         if not kwargs:
             kwargs["category"] = None
-        return self.get(slug__iexact=component, project__slug__iexact=project, **kwargs)
+        return self.filter(
+            slug__iexact=component, project__slug__iexact=project, **kwargs
+        )
+
+    def get_by_path(self, path: str) -> Component:
+        return self.filter_by_path(path).get()
+
+    def get_linked(self, val):
+        """Return component for linked repo."""
+        if not is_repo_link(val):
+            return None
+        return self.get_by_path(val[10:])
 
     def order_project(self):
         """Ordering in global scope by project name."""
@@ -973,7 +984,7 @@ class Component(models.Model, PathMixin, CacheKeyMixin, ComponentCategoryMixin):
                 continue
 
             if addon.has_settings():
-                form = addon.get_add_form(None, component, data=configuration)
+                form = addon.get_add_form(None, component=component, data=configuration)
                 if not form.is_valid():
                     component.log_warning(
                         "could not enable addon %s, invalid settings", name
@@ -986,7 +997,7 @@ class Component(models.Model, PathMixin, CacheKeyMixin, ComponentCategoryMixin):
 
             component.log_info("enabling addon %s", name)
             # Running is disabled now, it is triggered in after_save
-            addon.create(component, run=False, configuration=configuration)
+            addon.create(component=component, run=False, configuration=configuration)
 
     def create_glossary(self) -> None:
         project = self.project
@@ -3571,7 +3582,7 @@ class Component(models.Model, PathMixin, CacheKeyMixin, ComponentCategoryMixin):
 
         result = defaultdict(list)
         result["__lookup__"] = {}
-        for addon in Addon.objects.filter_component(self):
+        for addon in Addon.objects.filter_for_execution(self):
             for installed in addon.event_set.all():
                 result[installed.event].append(addon)
             result["__all__"].append(addon)